backend/pyproject.toml
[project]
name = "enterprise-rag-backend"
version = "0.1.0"
description = "Production-ready RAG with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # FastAPI Core
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",

    # Database
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",

    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "fastapi-limiter>=0.1.6",

    # Queue & Worker
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",

    # Vector DB & Embeddings
    "pinecone-client>=5.0.0",
    "google-generativeai>=0.8.0",

    # Document Processing
    "unstructured[pdf]>=0.16.0",
    "llama-index>=0.11.0",
    "llama-index-vector-stores-pinecone>=0.2.0",
    "llama-index-embeddings-gemini>=0.1.0",

    # Reranking
    "cohere>=5.13.0",
    "sentence-transformers>=3.3.0",

    # S3/Storage
    "boto3>=1.35.0",
    "minio>=7.2.0",

    # Resilience
    "tenacity>=9.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",

    # Code Quality
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "black>=24.10.0",
    "isort>=5.13.0",

    # Evaluation
    "ragas>=0.2.0",

    # Documentation
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
]

# --- THIS SECTION FIXES THE BUILD ERROR ---
[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
